Move NDCubeSequence to a seperate file
